aider: Add unit test for fizzbuzz endpoint.
# Aider chat conversation:

USER: create a unit test for fizzbuzz endpoint

ASSISTANT: To create a unit test for the `/fizzbuzz` endpoint in the Flask app, we need to modify the `flask_app/test_app.py` file. Here's what we'll do:

1. Add a new test method to the `FlaskAppTestCase` class for the `/fizzbuzz` endpoint.
2. Use the test client to make a GET request to the `/fizzbuzz` endpoint.
3. Assert that the status code of the response is 200 (OK).
4. Assert that the response data is correct. Since the response is JSON, we'll need to parse it and check that it contains the expected FizzBuzz sequence.

Here is the *SEARCH/REPLACE block* to add the unit test for the FizzBuzz endpoint:

flask_app/test_app.py
<source>python
<<<<<<< SEARCH
if __name__ == '__main__':
    unittest.main()
=======
    def test_fizzbuzz_endpoint(self):
        response = self.client.get('/fizzbuzz')
        self.assertEqual(response.status_code, 200)
        fizzbuzz_results = response.json['results']
        expected_results = []
        for i in range(1, 101):
            if i % 3 == 0 and i % 5 == 0:
                expected_results.append('FizzBuzz')
            elif i % 3 == 0:
                expected_results.append('Fizz')
            elif i % 5 == 0:
                expected_results.append('Buzz')
            else:
                expected_results.append(str(i))
        self.assertEqual(fizzbuzz_results, expected_results)

if __name__ == '__main__':
    unittest.main()
>>>>>>> REPLACE
</source>

This code will add a new test method `test_fizzbuzz_endpoint` that checks if the `/fizzbuzz` endpoint is functioning correctly by comparing the returned JSON results with the expected FizzBuzz sequence.